+++ components/Admin/Paper/PaperNewPage.tsx
@@ -2,34 +2,15 @@ import useInput from '../../../hooks/useInput';
 import { postPaper } from '../../../api/paper';
 import Input from '../../Element/Shared/Input';
 import AdminButton from '../../Element/Admin/AdminButton';
-import { useState } from 'react';
 import { SC } from '../../../styles/styled';
+import useKeywords from '../../../hooks/useKeywords';
 
 export default function PaperNewPage() {
 	const { value: title, onChange: onChangeTitle } = useInput();
 	const { value: year, onChange: onChangeYear } = useInput();
 	const { value: researcherName, onChange: onChangeResearcherName } = useInput();
 	const { value: doi, onChange: onChangeDoi } = useInput();
-	const [keywords, setKeywords] = useState<string[]>(['']);
-
-	const onChangeKeyword = (e: React.ChangeEvent<HTMLInputElement>, idx: number) => {
-		const copy = [...keywords];
-		copy[idx] = e.target.value;
-		setKeywords(copy);
-	};
-
-	const onAddKeyword = () => {
-		const copy = [...keywords];
-		copy.push('');
-		setKeywords(copy);
-	};
-
-	const onRemoveKeyword = (idx: number) => {
-		const copy = [...keywords];
-		copy.splice(idx, 1);
-		console.log(copy);
-		setKeywords(copy);
-	};
+	const { keywords, onChangeKeyword, onAddKeyword, onRemoveKeyword } = useKeywords();
 
 	const onSubmit = async () => {
 		/* POST */
